requirements.txt
# Core Data Processing
pandas>=2.2.0
numpy>=1.26.0
pytz>=2023.3

# NLP - Hugging Face
transformers>=4.36.0
torch>=2.2.0
sentencepiece>=0.1.99
accelerate>=0.25.0

# NLP - Language Detection
langdetect>=1.0.9

# PII Detection
presidio-analyzer>=2.2.35
presidio-anonymizer>=2.2.35
spacy>=3.7.2

# Topic Modeling (v2 feature)
bertopic>=0.16.0
sentence-transformers>=2.2.2
umap-learn>=0.5.5
hdbscan>=0.8.33

# Text Processing & ML
scikit-learn>=1.3.2
nltk>=3.8.1

# Progress Bars
tqdm>=4.66.1

# API & Network
requests>=2.31.0
urllib3>=2.1.0

# Visualization
wordcloud>=1.9.3
matplotlib>=3.8.2
Pillow>=10.1.0

# Environment & Config
python-dotenv>=1.0.0

# GUI - Menu Bar App (v2 feature)
rumps>=0.4.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# Type Checking
mypy>=1.7.1
